create CurrentLocation component
